requirements.txt
python
Django
djangorestframework
